refactor: coloca middlewares de User em uma pasta User
